Crée des sous-dossiers pour la gestion de contributeurs
On crée :
 `/invitation`
 `/kit`
 `/personnalisation`
 `/suppression`
